refactor(spaces): separate link handling from bootstrap dependent code
